This commit adds three scripts:
1. "csv-txt_converter.py" is a script which takes a csv files with
headings "date", "heading", "body", and turns each row into a .txt file
using the date as the file name and adding the heading and body to the
txt file.
2. "herald_scraper_2.py" is a web scraper which scrapes freedom related
covid-19 articles from the NZ herald and outputs a .txt for each
article.
3. "RNZ_scraper.py" is a script which is similaar to "news_scraper.py"
but outputs a single .txt for each article scraped instead of a single
csv.